PCLOUDS-1602: SFM refactor - extract dashboards code
